+++ .idea/misc.xml
@@ -1,18 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
-  <component name="DesignSurface">
-    <option name="filePathToZoomLevelMap">
-      <map>
-        <entry key="..\:/Users/asus/StudioProjects/onppe_v1/app/src/main/res/drawable/custom_edittext.xml" value="0.164" />
-        <entry key="..\:/Users/asus/StudioProjects/onppe_v1/app/src/main/res/drawable/custom_edittext2.xml" value="0.164" />
-        <entry key="..\:/Users/asus/StudioProjects/onppe_v1/app/src/main/res/drawable/custom_rectangle.xml" value="0.164" />
-        <entry key="..\:/Users/asus/StudioProjects/onppe_v1/app/src/main/res/drawable/ic_baseline_calendar_today_24.xml" value="0.164" />
-        <entry key="..\:/Users/asus/StudioProjects/onppe_v1/app/src/main/res/layout/fragment_signalement_form1.xml" value="0.4" />
-        <entry key="..\:/Users/asus/StudioProjects/onppe_v1/app/src/main/res/layout/fragment_signalement_form2.xml" value="0.2" />
-        <entry key="..\:/Users/asus/StudioProjects/onppe_v1/app/src/main/res/layout/fragment_signalement_form3.xml" value="0.264" />
-      </map>
-    </option>
-  </component>
   <component name="ExternalStorageConfigurationManager" enabled="true" />
   <component name="ProjectRootManager" version="2" languageLevel="JDK_11" default="true" project-jdk-name="Android Studio default JDK" project-jdk-type="JavaSDK">
     <output url="file://$PROJECT_DIR$/build/classes" />
