--- app/src/main/res/drawable/custom_edittext.xml
@@ -1,9 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<shape xmlns:android="http://schemas.android.com/apk/res/android"
-    android:shape="rectangle">
-    <solid android:color="@color/white"/>
-    <corners  android:radius="10dp"/>
-    <stroke
-        android:width="1dp"
-        android:color="#BEB7B5"/>
-</shape>
