--- app/src/main/res/drawable/custom_rectangle.xml
@@ -1,6 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<shape xmlns:android="http://schemas.android.com/apk/res/android"
-    android:shape="rectangle">
-    <solid android:color="@color/green"/>
-    <corners android:bottomRightRadius="20dp" android:bottomLeftRadius="20dp"/>
-</shape>
