+++ app/src/main/res/values/colors.xml
@@ -7,5 +7,4 @@
     <color name="teal_700">#FF018786</color>
     <color name="black">#FF000000</color>
     <color name="white">#FFFFFFFF</color>
-    <color name="green">#59C55E</color>
 </resources>
